True Gacha Mode Implemented
Implemented the new "True Gacha Mode", which can be enabled or disabled to change the rates of pulls in the game. When enabled, the rates/chances for each rank will be much closer (i.e. more unfair) to real gacha games. Otherwise, the rates will stay the same as they have been.